+++ tests/testthat/test-variable-catalog.R
@@ -76,6 +76,11 @@ with(fake.HTTP, {
             c("Gender", "Birth Year", "starttime"))
         expect_identical(aliases(varcat)[1:2], c("gender", "birthyr"))
     })
+    
+    test_that("types getter", {
+        expect_identical(types(varcat)[1:3],
+            c("categorical", "numeric", "datetime"))
+    })
 })
 
 if (run.integration.tests) {